:ambulance: Fix avatar not being shown
With the addition of the avatar image to the vCard file, image handling
was moved away from classic `<img>` to `fetch`. It was done to be able
to read image data, encode it to vCard and wite it to vCard.

Problem wit fetch is CORS. If no CORS headers were set by the server
serving the image, the image could not be downloaded any more. This is
not a problem with `<img>`, but then we cannot access the images data
anymore.

So the solution here is to still use fetch. In case it works we can
write vCard data and show the avatar from fetch result. In case it does
not work, we at least show the avatar using `<img>` but won't write the
avatar to the vCard.